Reworked Calendar to Work with crewFlightController
Assign crew member also works, currently in order to see crew member assignments you click on the text on a day with a flight in the calendar, and it will print to the terminal. I'm thinking about having another list in the flights and crew members tabs that display the assignments.